refactor : domain 내부 vo 클래스 패키지 분류
