+++ src/main/java/roomescape/domain/Reservation.java
@@ -1,5 +1,8 @@
 package roomescape.domain;
 
+import roomescape.domain.vo.ReservationDate;
+import roomescape.domain.vo.ReservationTime;
+
 public class Reservation {
 
     private final Long id;
