+++ src/main/java/roomescape/domain/ReservationTimeRepository.java
@@ -3,6 +3,7 @@ package roomescape.domain;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
+import roomescape.domain.vo.ReservationTime;
 
 public interface ReservationTimeRepository {
