+++ src/main/java/roomescape/domain/vo/MemberName.java
@@ -1,20 +1,21 @@
-package roomescape.domain;
+package roomescape.domain.vo;
 
 import io.micrometer.common.util.StringUtils;
 import java.util.Objects;
 
-public class MemberPassword {
+public class MemberName {
+    private static final int MAX_NAME_SIZE = 10;
 
     private final String value;
 
-    public MemberPassword(String value) {
+    public MemberName(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (StringUtils.isBlank(value)) {
-            throw new IllegalArgumentException("비밀번호는 필수입니다.");
+        if (StringUtils.isBlank(value) || value.length() > MAX_NAME_SIZE) {
+            throw new IllegalArgumentException(value + " 사용자 이름 1글자 이상 10글자 이하를 입력해주세요.");
         }
     }
 
@@ -23,14 +24,14 @@ public class MemberPassword {
     }
 
     @Override
-    public boolean equals(Object o) {
+    public boolean equals(final Object o) {
         if (this == o) {
             return true;
         }
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        MemberPassword that = (MemberPassword) o;
+        MemberName that = (MemberName) o;
         return Objects.equals(value, that.value);
     }
