+++ src/main/java/roomescape/domain/vo/MemberPassword.java
@@ -1,21 +1,20 @@
-package roomescape.domain;
+package roomescape.domain.vo;
 
 import io.micrometer.common.util.StringUtils;
 import java.util.Objects;
 
-public class MemberRole {
+public class MemberPassword {
 
-    private static final int MAX_ROLE_SIZE = 10;
     private final String value;
 
-    public MemberRole(String value) {
+    public MemberPassword(String value) {
         validate(value);
         this.value = value;
     }
 
     private void validate(String value) {
-        if (StringUtils.isBlank(value) || value.length() > MAX_ROLE_SIZE) {
-            throw new IllegalArgumentException("사용자의 권한은 1글자 이상 10글자 이하 입니다.");
+        if (StringUtils.isBlank(value)) {
+            throw new IllegalArgumentException("비밀번호는 필수입니다.");
         }
     }
 
@@ -24,14 +23,14 @@ public class MemberRole {
     }
 
     @Override
-    public boolean equals(final Object o) {
+    public boolean equals(Object o) {
         if (this == o) {
             return true;
         }
         if (o == null || getClass() != o.getClass()) {
             return false;
         }
-        MemberRole that = (MemberRole) o;
+        MemberPassword that = (MemberPassword) o;
         return Objects.equals(value, that.value);
     }
