+++ src/main/java/roomescape/infrastructure/JdbcReservationTimeRepositoryImpl.java
@@ -11,8 +11,8 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeRepository;
+import roomescape.domain.vo.ReservationTime;
 
 @Repository
 public class JdbcReservationTimeRepositoryImpl implements ReservationTimeRepository {
