+++ src/main/java/roomescape/service/MemberAuthService.java
@@ -4,11 +4,11 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import roomescape.domain.Member;
-import roomescape.domain.MemberEmail;
-import roomescape.domain.MemberName;
-import roomescape.domain.MemberPassword;
 import roomescape.domain.MemberRepository;
-import roomescape.domain.MemberRole;
+import roomescape.domain.vo.MemberEmail;
+import roomescape.domain.vo.MemberName;
+import roomescape.domain.vo.MemberPassword;
+import roomescape.domain.vo.MemberRole;
 import roomescape.exception.AuthorizationException;
 import roomescape.exception.EmailAlreadyExistsException;
 import roomescape.service.request.MemberAppRequest;
