+++ src/test/java/roomescape/domain/MemberPasswordTest.java
@@ -7,6 +7,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
+import roomescape.domain.vo.MemberPassword;
 
 class MemberPasswordTest {
