+++ src/test/java/roomescape/infrastructure/JdbcReservationTimeRepositoryImplTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.domain.ReservationTime;
 import roomescape.domain.ReservationTimeRepository;
+import roomescape.domain.vo.ReservationTime;
 
 @JdbcTest
 class JdbcReservationTimeRepositoryImplTest {
